requirements.txt
numpy
pillow
argparse